api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
python-multipart==0.0.20
requests==2.32.5
beautifulsoup4==4.13.4
arxiv==2.2.0
PyMuPDF==1.26.4
pdfplumber==0.11.7
openai==1.102.0
google-genai==1.68.0
chromadb==1.0.20
tiktoken==0.11.0
